Add input to dry run result (#10236)
## Description

This adds `input` to the dry run result, which wallets can use this to
easily deserialize input values, which can be hard to do otherwise. This
also means clients can easily get the parsed structure of a transaction
without needing to either execute the transaction or needing a full BCS
parser (neat!).

Two notes:
- I opted for `input` here, whereas on the execute transaction block
result it's `transaction`. The `transaction` terminology seemed out of
date, and this aligns closer with `showInput` on execute transaction
block.
- This doesn't return the same thing as the input on the
`executeTransactionBlock` because they take different inputs (signed
transaction vs just transaction data). It's not really possible to
reconcile this and I think this is fine in the sense that it still hands
back the parsed input.

## Test Plan

Added to TS SDK, tests should continue to pass.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes