[indirect objects] check presence via multi_get_raw_bytes api (#10377)
When the reference count of the indirect move object goes to zero, the
value in rocksdb might still be present for some time before two
consecutive compaction jobs clear it out.
This is expected and represents a valid state of the system. 
However, for the `get` method, the return value will be merged into the
reference count on read. The call site for `update_objects_and_locks`
performs a read-before-write operation and needs to be switched to the
`multi_get_raw_bytes` API instead because the squashed reference count
value cannot be deserialized as a `StoreMoveObject`.

Furthermore, if the database value is indeed a reference count, we need
to trigger a full merge for the object instead of a partial merge.
That's because we can't differentiate between two system states:
* the value is merged to a zero reference count, but the original full
merge operand is still present
* the value is merged to zero, but the original operand was erased
during the first round of compaction