Increase number of pending certificates buffered at synchronizer (#11000)
## Description 

We have observed nodes being a few hundred rounds behind and having a
hard time catching up. Increase the number of rounds of certificates to
buffer, which should still be acceptable in memory usage, and hopefully
help with catching up when the node is only a few hundred rounds behind.

## Test Plan 

Deploy to private testnet

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes