Add caching for account address balance computation (#10693)
## Description 

Add a cache which caches the account balance and invalidates it whenever
coin ownership changes. This prevents background db access when coin
ownership has not changed. Most of the interesting changes are in
`indexes.rs`. Only when `deleted` and `new` coins for an account are
detected, we invalidate the caches.

## Test Plan 

Added tests to ensure caches are returning latest data as expected.