Make coin metadata nullable (#11047)
## Description 

There are cases where coin metadata fails to load, and these are
actually somewhat expected. Rather than making the API hard error on
these case, we instead can just make it return `null`. In general I'd
like it if more of our APIs did this so that errors were really an
actual error vs just "there is no data".

This is technically a breaking change, but I opted to not introduce it
gradually because I think this should be fairly non-impacting to the
ecosystem

## Test Plan 

Ran locally.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes