[sui-verifier] Move helper functions into more foundational crates (#11135)
## Description 

`sui-verifier` contained a number of helper functions that were being
used by other crates (like `sui-json`). These helpers don't need to be
versioned as part of the execution layer, but `sui-verifier` in general
does. By moving helpers into `sui-types` and `move-bytecode-utils`
(which are not versioned), we avoid `sui-json` having to deal with the
versioned API.

## Test Plan 

```
$ cargo simtest
$ env SUI_SKIP_SIMTESTS=1 cargo nextest run
```